Add ValueHolderCodec
