+++ airframe-opts/src/main/scala/wvlet/airframe/opts/HelpMessage.scala
@@ -12,6 +12,7 @@
  * limitations under the License.
  */
 package wvlet.airframe.opts
+import wvlet.airframe.opts.OptionParser.{CLArgItem, CLOption}
 
 /**
   *
@@ -22,6 +23,87 @@ class HelpMessage {
     s"""|usage:${commandName} ${argumentList}
         |  ${description}
         |${optionList}""".stripMargin
+  }
+
+  protected def defaultUsage(args: Seq[CLArgItem]): String = {
+    val l = for (a <- args) yield {
+      a.name
+    }
+    l.map("[%s]".format(_)).mkString(" ")
+  }
+
+  def printHelp(stack: List[LauncherInstance] = Nil): Unit = {
+    trace("print usage")
+    val p = optionParser
+    printUsage
+
+    // Show parent options
+    val parentOptions = stack.flatMap { x =>
+      x.launcher.optionList
+    }
+    if (parentOptions.nonEmpty) {
+      println("[global options]")
+      println(parentOptions.mkString("\n"))
+    }
+
+    if (subCommands.nonEmpty) {
+      println("[commands]")
+
+      val maxCommandNameLen = subCommands.map(_.name.length).max
+      val format            = " %%-%ds\t%%s".format(math.max(10, maxCommandNameLen))
+      // Show sub commend lists
+      subCommands.foreach { c =>
+        println(format.format(c.name, c.description))
+      }
+    }
+  }
+
+  def createOptionHelpMessage = {
+    val optionList = createOptionList
+    val b          = new StringBuilder
+    if (optionList.nonEmpty) {
+      b.append("[options]\n")
+      b.append(optionList.mkString("\n") + "\n")
+    }
+    b.result
+  }
+
+  def createOptionList: Seq[String] = {
+    val optDscr: Seq[(CLOption, String)] = for (o <- optionList) yield {
+      val prefixes = o.prefixes
+      val hasShort = prefixes.exists(_.length == 2)
+      val hasAlias = prefixes.exists(_.length > 2)
+      val l        = new StringBuilder
+      l.append(prefixes.mkString(", "))
+
+      if (o.takesArgument) {
+        if (hasAlias) {
+          l append ":"
+        } else if (hasShort) {
+          l append " "
+        }
+        l append "[%s]".format(o.param.name.toUpperCase)
+      }
+      (o, l.toString)
+    }
+
+    val optDscrLenMax =
+      if (optDscr.isEmpty) {
+        0
+      } else {
+        optDscr.map(_._2.length).max
+      }
+
+    def genDescription(opt: CLOption) = {
+      opt.annot.description()
+    }
+
+    val s = for (x <- optDscr) yield {
+      val paddingLen = optDscrLenMax - x._2.length
+      val padding    = Array.fill(paddingLen)(" ").mkString
+      " %s%s  %s".format(x._2, padding, genDescription(x._1))
+    }
+    s
   }
 
 }
